+++ src/main/java/com/zoopick/server/dto/match/ItemMatchResultResponse.java
@@ -1,4 +1,4 @@
-package com.zoopick.server.dto.item;
+package com.zoopick.server.dto.match;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.zoopick.server.entity.MatchStatus;
